nl_ingest/requirements.txt
geopy==2.4.1
flask-sqlalchemy==3.1.1
feedparser==6.0.11
dotenv==0.9.9
psycopg2-binary==2.9.10
ratelimit==2.2.1
python-dateutil==2.9.0
spacy==3.8.7
